requirements.txt
click==8.0.4
Faker==12.0.0
psycopg2-binary==2.9.3
pypika==0.48.8
requests==2.27.1
schedule==1.1.0
xmltodict==0.12.0
